+++ CHANGELOG.md
@@ -3,12 +3,16 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [1.1.1] - 2020-07-03
+- Added withPayIdFormat() method.
+- Added withoutMoneyCast() method.
+- Added `pay_id` field in panel views.
 
 ## [1.1.0] - 2020-06-16
-- Added extra amounts in payments (for taxes, shipping...)
+- Added extra amounts in payments (for taxes, shipping...).
 
 ## [1.0.1] - 2020-06-04
-- Fixed issues
+- Fixed issues.
 
 ## [1.0.0] - 2020-06-04
-- Initial release
+- Initial release.